move is_dtype_supported to test.helpers
updated all places that check if float16 is supports
