+++ test/external/external_test_onnx_backend.py
@@ -3,9 +3,9 @@ from typing import Any, Tuple
 from onnx.backend.base import Backend, BackendRep
 import onnx.backend.test
 import numpy as np
-from tinygrad.tensor import Tensor
-from tinygrad.helpers import getenv, CI, OSX
-from tinygrad.device import Device
+from tinygrad import Tensor, Device, dtypes
+from tinygrad.helpers import getenv, OSX
+from test.helpers import is_dtype_supported
 
 # pip3 install tabulate
 pytest_plugins = 'onnx.backend.test.report',
@@ -49,7 +49,7 @@ backend_test.exclude('test_adam_multiple_cpu')
 backend_test.exclude('test_nesterov_momentum_cpu')
 
 # about different dtypes
-if Device.DEFAULT in ["METAL"] or (OSX and Device.DEFAULT == "GPU"):
+if not is_dtype_supported(dtypes.float64):
   backend_test.exclude('float64')
   backend_test.exclude('DOUBLE')
   # these have float64 inputs
@@ -59,8 +59,7 @@ if Device.DEFAULT in ["METAL"] or (OSX and Device.DEFAULT == "GPU"):
   backend_test.exclude('test_einsum_*')
   backend_test.exclude('test_cumsum_*')
 
-# no float16 in CI, LLVM segfaults, GPU requires cl_khr_fp16
-if Device.DEFAULT in ['LLVM', 'CUDA', 'GPU'] and CI:
+if not is_dtype_supported(dtypes.float16):
   backend_test.exclude('float16')
   backend_test.exclude('FLOAT16')
